put components into files, remove some builtin stuff
