+++ src/components/AppError.js
@@ -1,0 +1,7 @@
+export default function AppError(props) {
+  return (
+    <div className="error">
+      Something happened: { props.error }
+    </div>
+  );
+}
