+++ src/components/AppLoading.js
@@ -1,0 +1,7 @@
+export default function AppLoading(props) {
+  return (
+    <div className="loading">
+      Loading apps
+    </div>
+  )
+}
